Move the matchedDepsDiffTable function to a more global utils.ts
